Class providing some methods to store data into a hashtable and access stored items via key

R lists with named elements are not hashed. Hash lookups are O(1), because during insert the key is converted to an integer using a hash function, and then the value put in the space hash(key) (this is a big simplification and avoids the complexity of dealing with collisions). Lookups of the key just require hashing the key to find the value's position (which is O(1), versus a O(n) array lookup). R lists use name lookups which are O(n).

HashMap

Format

R6Class object.

Value

Object of R6Class with methods for key-value mapping

Fields

field(s)

access all via ()

Construction

``` HashMap$new() ```

Methods

Documentation

For full documentation of each method go to https://github.com/ChrisNice89/ORM

new()

This method creates a HashMap instance.

put(key = "", value = NA)

store key-value pair(s)

get(key,unname=TRUE)

access item(s) with associated key(s)

See also

Examples

hm<-HashMap$new()

#> <Validator> for parent class: <HashMap> created

keys<- c("tic", "tac", "toe")
values <- c(1, 22, 333)

hm$put("key",1)
hm$put(key = keys,value = values)
hm$get(keys)

#> [1]   1  22 333

hm$get("tic")

#> [1] 1

hm$size()

#> [1] 4

hm

#> <HashMap> with name: <Default Dictionary> created
#> 	<key-value-pairs> 	
#>      key  ::  1
#>      tac  ::  22
#>      tic  ::  1
#>      toe  ::  333

hm$remove("tic")
hm$get("tic")

#> Error in (function (x, pos = -1L, envir = as.environment(pos), mode = "any",     inherits = TRUE) .Internal(get(x, envir, mode, inherits)))(x = dots[[1L]][[1L]],     pos = <environment>): Objekt 'tic' nicht gefunden

hm$size()

#> [1] 3

hm$getKeys()[1]

#> [1] "key"

hm$getValues()

#> Error in return(setNames(private$.hash.get(ls(hash), private$.hash)),     NULL): mehrargumentige Rückgaben sind nicht erlaubt